Moved emulator setup into own class
