+++ emulator/src/tests/helpers/CaptureSerial.hpp
